+++ JHazm/main/java/iust/ac/ir/nlp/jhazm/BijankhanReader.java
@@ -1,20 +1,13 @@
-package JHazm;
+package iust.ac.ir.nlp.jhazm;
 
 import com.infomancers.collections.yield.Yielder;
 import edu.stanford.nlp.ling.TaggedWord;
-import java.io.BufferedReader;
-import java.io.DataInputStream;
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.io.InputStreamReader;
+
+import java.io.*;
 import java.nio.charset.Charset;
 import java.nio.file.Files;
 import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
+import java.util.*;
 
 /**
  * interfaces Bijankhan Corpus (http://ece.ut.ac.ir/dbrg/bijankhan/Corpus/BijanKhan_Corpus_Processed.zip) that 
@@ -26,16 +19,45 @@ public class BijankhanReader {
     private final String[] punctuation = new String[] { "#", "*", ".", "؟", "!" };
 
     private String bijankhanFile;
+    private boolean joinedVerbParts;
+    private String posMap;
+    private Normalizer normalizer;
+    private WordTokenizer tokenizer;
+
+    public BijankhanReader() throws IOException {
+        this("resources/bijankhan.txt", true, "data/posMaps.dat");
+    }
+
+    public BijankhanReader(boolean joinedVerbParts) throws IOException {
+        this("resources/bijankhan.txt", joinedVerbParts, "data/posMaps.dat");
+    }
+
+    public BijankhanReader(String posMap) throws IOException {
+        this("resources/bijankhan.txt", true, posMap);
+    }
+
+    public BijankhanReader(boolean joinedVerbParts, String posMap)
+            throws IOException {
+        this("resources/bijankhan.txt", joinedVerbParts, posMap);
+    }
+
+    public BijankhanReader(String bijankhanFile, boolean joinedVerbParts, String posMap)
+            throws IOException {
+        this.bijankhanFile = bijankhanFile;
+        this.joinedVerbParts = joinedVerbParts;
+        this.posMap = posMap;
+        this.normalizer = new Normalizer(true, false, true);
+        this.tokenizer = new WordTokenizer();
+    }
+
     public String getBijankhanFile() {
         return bijankhanFile;
     }
-    
-    private boolean joinedVerbParts;
+
     public boolean isJoinedVerbParts() {
         return joinedVerbParts;
     }
-    
-    private String posMap;
+
     private HashMap GetPosMap() throws IOException
     {
         if (this.posMap != null)
@@ -53,43 +75,35 @@ public class BijankhanReader {
             return null;
     }
 
-    private Normalizer normalizer;
     public Normalizer getNormalizer() {
         return normalizer;
-    }
-    
-    private WordTokenizer tokenizer;
-
-    public BijankhanReader() throws IOException {
-        this("resources/bijankhan.txt", true, "data/posMaps.dat");
-    }
-
-    public BijankhanReader(boolean joinedVerbParts) throws IOException {
-        this("resources/bijankhan.txt", joinedVerbParts, "data/posMaps.dat");
-    }
-
-    public BijankhanReader(String posMap) throws IOException { 
-        this("resources/bijankhan.txt", true, posMap);
-    }
-
-    public BijankhanReader(boolean joinedVerbParts, String posMap) 
-            throws IOException {
-        this("resources/bijankhan.txt", joinedVerbParts, posMap);
-    }
-
-    public BijankhanReader(String bijankhanFile, boolean joinedVerbParts, String posMap) 
-            throws IOException {
-        this.bijankhanFile = bijankhanFile;
-        this.joinedVerbParts = joinedVerbParts;
-        this.posMap = posMap;
-        this.normalizer = new Normalizer(true, false, true);
-        this.tokenizer = new WordTokenizer();
     }
 
     public Iterable<List<TaggedWord>> GetSentences() {
         return new YieldSentence();
     }
-    
+
+    private List<TaggedWord> JoinVerbParts(List<TaggedWord> sentence) {
+        Collections.reverse(sentence);
+        List<TaggedWord> result = new ArrayList<>();
+        TaggedWord beforeTaggedWord = new TaggedWord("", "");
+        for (TaggedWord taggedWord : sentence) {
+            if (this.tokenizer.getBeforeVerbs().contains(taggedWord.word()) ||
+                    (this.tokenizer.getAfterVerbs().contains(beforeTaggedWord.word()) &&
+                            this.tokenizer.getVerbs().contains(taggedWord.word()))) {
+                beforeTaggedWord.setWord(taggedWord.word() + " " + beforeTaggedWord.word());
+                if (result.isEmpty())
+                    result.add(beforeTaggedWord);
+            } else {
+                result.add(taggedWord);
+                beforeTaggedWord = taggedWord;
+            }
+        }
+
+        Collections.reverse(result);
+        return result;
+    }
+
     class YieldSentence extends Yielder<List<TaggedWord>> {
         private BufferedReader br;
 
@@ -109,9 +123,9 @@ public class BijankhanReader {
             try {
                 HashMap mapper = GetPosMap();
                 List<TaggedWord> sentence = new ArrayList<>();
-                
+
                 String line;
-                
+
                 while ((line = br.readLine()) != null) {
                     String[] parts = line.trim().split("  +");
                     if (parts.length == 2) {
@@ -130,7 +144,7 @@ public class BijankhanReader {
 
                                 if (mapper != null) {
                                     for (TaggedWord tword : sentence) {
-                                        tword.setTag(mapper.get(tword.tag()).toString()); 
+                                        tword.setTag(mapper.get(tword.tag()).toString());
                                     }
                                 }
 
@@ -145,27 +159,5 @@ public class BijankhanReader {
                 ex.printStackTrace();
             }
         }
-    }
-
-    private List<TaggedWord> JoinVerbParts(List<TaggedWord> sentence) {
-        Collections.reverse(sentence);
-        List<TaggedWord> result = new ArrayList<>();
-        TaggedWord beforeTaggedWord = new TaggedWord("", "");
-        for (TaggedWord taggedWord : sentence) {
-            if (this.tokenizer.getBeforeVerbs().contains(taggedWord.word()) ||
-                (this.tokenizer.getAfterVerbs().contains(beforeTaggedWord.word()) &&
-                 this.tokenizer.getVerbs().contains(taggedWord.word()))) {
-                beforeTaggedWord.setWord(taggedWord.word() + " " + beforeTaggedWord.word());
-                if (result.isEmpty())
-                    result.add(beforeTaggedWord);
-            }
-            else {
-                result.add(taggedWord);
-                beforeTaggedWord = taggedWord;
-            }
-        }
-
-        Collections.reverse(result);
-        return result;
     }
 }
