+++ JHazm/main/java/iust/ac/ir/nlp/jhazm/DependencyParser.java
@@ -1,81 +1,32 @@
-package JHazm;
+package iust.ac.ir.nlp.jhazm;
 
 import com.infomancers.collections.yield.Yielder;
 import edu.stanford.nlp.ling.TaggedWord;
+import org.maltparser.concurrent.ConcurrentMaltParserModel;
+import org.maltparser.concurrent.ConcurrentMaltParserService;
+import org.maltparser.concurrent.graph.ConcurrentDependencyGraph;
+import org.maltparser.core.exception.MaltChainedException;
+
 import java.io.File;
 import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.List;
-import org.maltparser.concurrent.ConcurrentMaltParserModel;
-import org.maltparser.concurrent.ConcurrentMaltParserService;
-import org.maltparser.concurrent.graph.ConcurrentDependencyGraph;
-import org.maltparser.core.exception.MaltChainedException;
 
 /**
  *
  * @author Mojtaba Khallash
  */
 public class DependencyParser {
-    private SentenceTokenizer sentenceTokenizer;
-    public SentenceTokenizer getSentenceTokenizer() {
-        if (sentenceTokenizer == null)
-            sentenceTokenizer = new SentenceTokenizer();
-        return sentenceTokenizer;
-    }
-    public void setSentenceTokenizer(SentenceTokenizer value) {
-        this.sentenceTokenizer = value;
-    }
-
-    private WordTokenizer wordTokenizer;
-    public WordTokenizer getWordTokenizer() throws IOException {
-        if (wordTokenizer == null)
-            wordTokenizer = new WordTokenizer();
-        return wordTokenizer;
-    }
-    public void setWordTokenizer(WordTokenizer value) {
-        this.wordTokenizer = value;
-    }
-
-    private Normalizer normalizer;
-    public Normalizer getNormalizer() {
-        return normalizer;
-    }
-    public void setNormalizer(Normalizer normalizer) {
-        this.normalizer = normalizer;
-    }
-
-    private Lemmatizer lemmatizer;
-    public Lemmatizer getLemmatizer() {
-        return lemmatizer;
-    }
-    public void setLemmatizer(Lemmatizer lemmatizer) {
-        this.lemmatizer = lemmatizer;
-    }
-    
     public POSTagger tagger;
-    public POSTagger getTagger() {
-        if (tagger == null)
-            tagger = new POSTagger();
-        return tagger;
-    }
-    public void setTagger(POSTagger value) {
-        this.tagger = value;
-    }
-    
+    private SentenceTokenizer sentenceTokenizer;
+    private WordTokenizer wordTokenizer;
+    private Normalizer normalizer;
+    private Lemmatizer lemmatizer;
     private String modelFile;
     private ConcurrentMaltParserModel model;
-    private ConcurrentMaltParserModel getModel() 
-            throws MalformedURLException, 
-                   MaltChainedException {
-        if (model == null) {
-            URL maltModelURL = new File(this.modelFile).toURI().toURL();
-            this.model = ConcurrentMaltParserService.initializeParserModel(maltModelURL);
-        }
-        return model;
-    }
 
-    public DependencyParser() { 
+    public DependencyParser() {
         this(null, null, "resources/langModel.mco");
     }
 
@@ -83,6 +34,62 @@ public class DependencyParser {
         this.tagger = tagger;
         this.lemmatizer = lemmatizer;
         this.modelFile = modelFile;
+    }
+
+    public SentenceTokenizer getSentenceTokenizer() {
+        if (sentenceTokenizer == null)
+            sentenceTokenizer = new SentenceTokenizer();
+        return sentenceTokenizer;
+    }
+
+    public void setSentenceTokenizer(SentenceTokenizer value) {
+        this.sentenceTokenizer = value;
+    }
+
+    public WordTokenizer getWordTokenizer() throws IOException {
+        if (wordTokenizer == null)
+            wordTokenizer = new WordTokenizer();
+        return wordTokenizer;
+    }
+
+    public void setWordTokenizer(WordTokenizer value) {
+        this.wordTokenizer = value;
+    }
+
+    public Normalizer getNormalizer() {
+        return normalizer;
+    }
+
+    public void setNormalizer(Normalizer normalizer) {
+        this.normalizer = normalizer;
+    }
+
+    public Lemmatizer getLemmatizer() {
+        return lemmatizer;
+    }
+
+    public void setLemmatizer(Lemmatizer lemmatizer) {
+        this.lemmatizer = lemmatizer;
+    }
+
+    public POSTagger getTagger() {
+        if (tagger == null)
+            tagger = new POSTagger();
+        return tagger;
+    }
+
+    public void setTagger(POSTagger value) {
+        this.tagger = value;
+    }
+
+    private ConcurrentMaltParserModel getModel()
+            throws MalformedURLException,
+                   MaltChainedException {
+        if (model == null) {
+            URL maltModelURL = new File(this.modelFile).toURI().toURL();
+            this.model = ConcurrentMaltParserService.initializeParserModel(maltModelURL);
+        }
+        return model;
     }
 
     // Gets list of raw text
@@ -98,35 +105,10 @@ public class DependencyParser {
             throws IOException {
         return new YieldParsedSentence(sentences);
     }
-    
-    class YieldParsedSentence extends Yielder<ConcurrentDependencyGraph> {
-        private final List<String> sentences;
-        private int index;
-        
-        public YieldParsedSentence(List<String> sentences) {
-            this.sentences = sentences;
-            index = -1;
-        }
-        
-        @Override
-        protected void yieldNextCore() {
-            try {
-                index++;
-                if (index < sentences.size()) {
-                    String sentence = sentences.get(index);
-                    List<String> words = getWordTokenizer().Tokenize(sentence);
-                    yieldReturn(RawParse(getTagger().BatchTag(words)));
-                }
-            }
-            catch (Exception ex) {
-                ex.printStackTrace();
-            }
-        }
-    }
-    
-    public ConcurrentDependencyGraph RawParse(List<TaggedWord> sentence) 
-            throws MalformedURLException, 
-                   MaltChainedException {
+
+    public ConcurrentDependencyGraph RawParse(List<TaggedWord> sentence)
+            throws MalformedURLException,
+            MaltChainedException {
         String[] conll = new String[sentence.size()];
         for (int i = 0; i < sentence.size(); i++) {
             TaggedWord taggedWord = sentence.get(i);
@@ -142,9 +124,34 @@ public class DependencyParser {
         return Parse(conll);
     }
 
-    public ConcurrentDependencyGraph Parse(String[] conllSentence) 
-            throws MalformedURLException, 
-                   MaltChainedException {
+    public ConcurrentDependencyGraph Parse(String[] conllSentence)
+            throws MalformedURLException,
+            MaltChainedException {
         return this.getModel().parse(conllSentence);
+    }
+
+    class YieldParsedSentence extends Yielder<ConcurrentDependencyGraph> {
+        private final List<String> sentences;
+        private int index;
+
+        public YieldParsedSentence(List<String> sentences) {
+            this.sentences = sentences;
+            index = -1;
+        }
+
+        @Override
+        protected void yieldNextCore() {
+            try {
+                index++;
+                if (index < sentences.size()) {
+                    String sentence = sentences.get(index);
+                    List<String> words = getWordTokenizer().Tokenize(sentence);
+                    yieldReturn(RawParse(getTagger().BatchTag(words)));
+                }
+            }
+            catch (Exception ex) {
+                ex.printStackTrace();
+            }
+        }
     }
 }
