+++ JHazm/main/java/iust/ac/ir/nlp/jhazm/Lemmatizer.java
@@ -1,14 +1,10 @@
-package JHazm;
+package iust.ac.ir.nlp.jhazm;
 
 import java.io.IOException;
 import java.nio.charset.Charset;
 import java.nio.file.Files;
 import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashSet;
-import java.util.HashMap;
-import java.util.List;
+import java.util.*;
 
 /**
  *
