+++ JHazm/main/java/iust/ac/ir/nlp/jhazm/utility/RegexPattern.java
@@ -1,17 +1,16 @@
-package JHazm.Utility;
+package iust.ac.ir.nlp.jhazm.utility;
 
 /**
  *
  * @author Mojtaba Khallash
  */
 public class RegexPattern {
+    private final String Pattern;
+    private final String Replace;
     public RegexPattern(String pattern, String replace) {
         this.Pattern = pattern;
         this.Replace = replace;
     }
-
-    private final String Pattern;
-    private final String Replace;
 
     public String Apply(String text) {
         return text.replaceAll(Pattern, Replace);
