+++ JHazm/test/java/iust/ac/ir/nlp/jhazm/Test/LemmatizerTest.java
@@ -1,6 +1,6 @@
-package java.JHazm.Test;
+package iust.ac.ir.nlp.jhazm.Test;
 
-import JHazm.Lemmatizer;
+import iust.ac.ir.nlp.jhazm.Lemmatizer;
 import org.junit.Test;
 
 import java.io.IOException;
